// extensions/be-discount-function/generated/api.ts
export type CartLinesDiscountsGenerateRunResult = {
  /** The list of operations to apply discounts to the cart. */
  operations: Array<CartOperation>;
};

/** The operations that can be performed to apply discounts to the cart. */
export type CartOperation =
  /**
   * An operation that selects which entered discount codes to accept. Use this to
   * validate discount codes from external systems.
   */
  { enteredDiscountCodesAccept: EnteredDiscountCodesAcceptOperation; enteredDiscountCodesReject?: never; orderDiscountsAdd?: never; productDiscountsAdd?: never; }
  |  /**
   * An operation that rejects entered discount codes with a custom message. Use
   * this to conditionally reject discount codes based on business logic. This
   * operation can only be used if the function is backed by an automatic discount.
   */
  { enteredDiscountCodesAccept?: never; enteredDiscountCodesReject: EnteredDiscountCodesRejectOperation; orderDiscountsAdd?: never; productDiscountsAdd?: never; }
  |  /** An operation that applies order discounts to a cart that share a selection strategy. */
  { enteredDiscountCodesAccept?: never; enteredDiscountCodesReject?: never; orderDiscountsAdd: OrderDiscountsAddOperation; productDiscountsAdd?: never; }
  |  /** An operation that applies product discounts to a cart that share a selection strategy. */
  { enteredDiscountCodesAccept?: never; enteredDiscountCodesReject?: never; orderDiscountsAdd?: never; productDiscountsAdd: ProductDiscountsAddOperation; };
// ...
export type EnteredDiscountCodesAcceptOperation = {
  /** The list of discount codes to accept. */
  codes: Array<DiscountCode>;
};

/**
 * An operation that rejects entered discount codes with a custom message. Use this
 * to conditionally reject discount codes based on business logic. This operation
 * can only be used if the function is backed by an automatic discount.
 */
export type EnteredDiscountCodesRejectOperation = {
  /** The list of discount codes to reject. */
  codes: Array<RejectedDiscountCode>;
  /** The custom message to display to the customer when the discount codes are rejected. */
  message: Scalars['String']['input'];
};
// ...
export type OrderDiscountsAddOperation = {
  /** The list of discounts that can be applied to an order. */
  candidates: Array<OrderDiscountCandidate>;
  /** The strategy that's applied to the list of discounts. */
  selectionStrategy: OrderDiscountSelectionStrategy;
};
// ...
export type ProductDiscountsAddOperation = {
  /** The list of products that are eligible for the discount. */
  candidates: Array<ProductDiscountCandidate>;
  /** The strategy that's applied to the list of products that are eligible for the cart line discount. */
  selectionStrategy: ProductDiscountSelectionStrategy;
};